Pull changes from homepage

+++ orgbee/src/app/pages/freedom-wall/freedom-wall.component.ts
@@ -1,12 +1,14 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { FormBuilder, FormGroup, Validators, AbstractControl } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { PostDialogComponent } from './post-dialog/post-dialog.component';
 import { DataService } from '../../../service/data.service';
 import { ToastrService } from 'ngx-toastr';
-
 import { Response } from '../../app.component';
 import { LoginService, UserInfo } from '../../../service/login.service';
+import { ConfirmationDialogService } from '../../../service/confirmation-dialog.service';
+import { SpinnerService } from '../../../service/spinner.service';
+import {NgxMasonryComponent}  from "ngx-masonry";
 
 export interface Post {
   subject: string;
@@ -14,10 +16,11 @@ export interface Post {
   background_color: string;
   showOptions?: boolean;
   id?: number; // for sample only
-  post_status_id:number; //Update to status
-  student_id:number;
-  deletion_req_count?:number;
+  post_status_id: number; //Update to status
+  student_id: number;
+  deletion_req_count?: number;
 }
+
 
 @Component({
   selector: 'app-freedom-wall',
@@ -26,6 +29,7 @@ export interface Post {
 })
 
 export class FreedomWallComponent implements OnInit {
+  @ViewChild(NgxMasonryComponent) masonry: NgxMasonryComponent;
   posts: Post[];
   showModal = false;
   selectedPost: Post;
@@ -38,14 +42,20 @@ export class FreedomWallComponent implements OnInit {
   postsPerPage = 4; 
   totalPages = 1; 
   showManageWallModal = false; 
-
+  deletePostCount = 0;
+  manageWallCount = 0;
+  showFilterMessage: boolean = false;
+  showSpinner = false;
   userInfo: UserInfo;
   requestDelete:Post[];
   constructor(
     private dialog: MatDialog, 
     private dataService: DataService, 
     private toastr: ToastrService,
-    private fb: FormBuilder, private loginService: LoginService) { }
+    private fb: FormBuilder, 
+    private loginService: LoginService,
+    private confirmationDialogService: ConfirmationDialogService,
+    private spinnerService: SpinnerService) { }
     
   ngOnInit(): void {
     this.showPosts();
@@ -63,8 +73,23 @@ export class FreedomWallComponent implements OnInit {
     this.loadPendingPosts();
     this.updatePaginatedPosts();
     this.getDeletionRequests();
+    
   }
-
+  reloadMasonryLayout() {
+    if (this.masonry) {
+      this.masonry.reloadItems(); 
+      this.masonry.layout(); 
+      
+    }
+  }
+  
+  isCurrentUserPost(post: Post): boolean {
+    console.log('Post:', post);
+    console.log('Current User ID:', this.userInfo.id);
+    return post.student_id === this.userInfo.id;
+  }
+  
+  
   updateTitleCharacterCount(): void {
     const subjectControl = this.freedomwallForm.get('newPostTitle');
     if (subjectControl && subjectControl.value.length > 30) {
@@ -88,7 +113,9 @@ export class FreedomWallComponent implements OnInit {
   }
 
   addPost() {
-    if (this.freedomwallForm.invalid) {
+    this.confirmationDialogService.confirmAction('Post Confirmation', 'Your post will be reviewed first by the officers. Are you sure you want to post this?', () => {
+    this.spinnerService.show('Submitting post...')
+      if (this.freedomwallForm.invalid) {
       this.freedomwallForm.markAllAsTouched();
       return;
     }
@@ -103,6 +130,8 @@ export class FreedomWallComponent implements OnInit {
     
     this.dataService.addPosts(newPost).subscribe((res: Response) => {
       this.response = res;
+            setTimeout(() => {
+        this.spinnerService.hide();
       if (this.response.code === 200) {
         this.toastr.success(JSON.stringify(this.response.message), '', {
           timeOut: 2000,
@@ -116,18 +145,23 @@ export class FreedomWallComponent implements OnInit {
           toastClass: 'custom-toast error'
         });
       }
-      this.showPosts();
+      if (this.showFilterMessage) {
+        this.filterPostsByUser();
+      } else {
+        this.showPosts();
+      }
       this.loadPendingPosts();
-    })
+    }, 500);
+  });
 
-    this.freedomwallForm.reset();
-    this.toggleModal();
-  }
+  this.freedomwallForm.reset();
+  this.toggleModal();
+});
+}
 
   showPosts() {
     this.dataService.getPosts().subscribe((posts: Post[]) => {
       this.posts = posts;
-      console.log(this.posts);
     });
   }
 
@@ -172,24 +206,37 @@ export class FreedomWallComponent implements OnInit {
   }
 
   deletePost(id: number) {
-    const post_id = { id: id };
-    this.dataService.deletePosts(post_id).subscribe((res: Response) => {
-      this.response = res;
-      if (this.response.code === 200) {
-        this.toastr.success(JSON.stringify(this.response.message), '', {
-          timeOut: 2000,
-          progressBar: true,
-          toastClass: 'custom-toast success'
-        });
-      } else {
-        this.toastr.error(JSON.stringify(this.response.message), '', {
-          timeOut: 2000,
-          progressBar: true,
-          toastClass: 'custom-toast error'
-        });
-      }
-      this.showPosts();
-      this.getDeletionRequests();
+    this.confirmationDialogService.confirmAction('Delete Confirmation', 'This action can\'t be undone. Are you sure you want to delete this post?', () => {
+      this.spinnerService.show('Deleting post...');
+      const post_id = { id: id };
+      this.dataService.deletePosts(post_id).subscribe((res: Response) => {
+        this.response = res;
+        setTimeout(() => {
+          this.spinnerService.hide();
+          if (this.response.code === 200) {
+            this.toastr.success(JSON.stringify(this.response.message), '', {
+              timeOut: 2000,
+              progressBar: true,
+              toastClass: 'custom-toast success'
+            });
+          } else {
+            this.toastr.error(JSON.stringify(this.response.message), '', {
+              timeOut: 2000,
+              progressBar: true,
+              toastClass: 'custom-toast error'
+            });
+          }
+          
+          // Update posts based on filter status
+          if (this.showFilterMessage) {
+            this.posts = this.posts.filter(post => post.id !== id);
+            this.reloadMasonryLayout();
+          } else {
+            this.showPosts();
+          }
+          this.getDeletionRequests();
+        }, 500);
+      });
     });
   }
 
@@ -207,6 +254,7 @@ export class FreedomWallComponent implements OnInit {
       this.totalPages = Math.ceil(this.pendingPosts.length / this.postsPerPage);
       this.currentPage = 1;
       this.updatePaginatedPosts();
+      this.updateButtonCounts();
     });
   }
 
@@ -230,48 +278,79 @@ export class FreedomWallComponent implements OnInit {
     }
   }
 
-  approvePost(postId: number) {
-    const post_id = { id: postId };
-    this.dataService.acceptPost(post_id).subscribe((res: Response)=>{
-      this.response =res;
-      if (this.response.code === 200) {
-        this.toastr.success(JSON.stringify(this.response.message), '', {
-          timeOut: 2000,
-          progressBar: true,
-          toastClass: 'custom-toast success'
-        });
-      } else {
-        this.toastr.error(JSON.stringify(this.response.message), '', {
-          timeOut: 2000,
-          progressBar: true,
-          toastClass: 'custom-toast error'
-        });
-      }
-      console.log(this.response);
-      this.showPosts();
-      this.loadPendingPosts();
+  approvePost(postId: number): void {
+    this.confirmationDialogService.confirmAction('Approve Confirmation', 'This action can\'t be undone. Are you sure you want to approve this post?', () => {
+      const post_id = { id: postId };
+      this.spinnerService.show('Approving post...');
+      this.dataService.acceptPost(post_id).subscribe(
+        (res: Response) => {
+          this.response = res;
+          setTimeout(() => {
+            this.spinnerService.hide();
+            if (this.response.code === 200) {
+              this.toastr.success(JSON.stringify(this.response.message), '', {
+                timeOut: 2000,
+                progressBar: true,
+                toastClass: 'custom-toast success'
+              });
+              // Refresh posts based on current view
+              if (this.showFilterMessage) {
+                this.showPosts();
+                this.filterPostsByUser();
+              } else {
+                this.showPosts();
+              }
+              this.loadPendingPosts(); // Refresh pending posts
+            } else {
+              this.toastr.error(JSON.stringify(this.response.message), '', {
+                timeOut: 2000,
+                progressBar: true,
+                toastClass: 'custom-toast error'
+              });
+            }
+          }, 500);
+        },
+        (error) => {
+          this.toastr.error('An error occurred while approving the post.', '', {
+            timeOut: 2000,
+            progressBar: true,
+            toastClass: 'custom-toast error'
+          });
+          console.error('Error approving post:', error);
+          this.spinnerService.hide();
+        }
+      );
     });
   }
-
+  
   declinePost(postId: number) {
-    const post_id = { id: postId };
-    this.dataService.declinePost(post_id).subscribe((res: Response)=>{
-      this.response =res;
-      if (this.response.code === 200) {
-        this.toastr.success(JSON.stringify(this.response.message), '', {
-          timeOut: 2000,
-          progressBar: true,
-          toastClass: 'custom-toast success'
-        });
-      } else {
-        this.toastr.error(JSON.stringify(this.response.message), '', {
-          timeOut: 2000,
-          progressBar: true,
-          toastClass: 'custom-toast error'
-        });
-      }
-      this.showPosts();
-      this.loadPendingPosts();
+    this.confirmationDialogService.confirmAction('Decline Confirmation', 'This action cant be undone. Are you sure you want to decline this post?', () => {
+      const post_id = { id: postId };
+      this.dataService.declinePost(post_id).subscribe((res: Response) => {
+        this.response = res;
+        setTimeout(() => {
+          if (this.response.code === 200) {
+            this.toastr.success(JSON.stringify(this.response.message), '', {
+              timeOut: 2000,
+              progressBar: true,
+              toastClass: 'custom-toast success'
+            });
+          } else {
+            this.toastr.error(JSON.stringify(this.response.message), '', {
+              timeOut: 2000,
+              progressBar: true,
+              toastClass: 'custom-toast error'
+            });
+          }
+  
+          if (this.showFilterMessage) {
+            this.filterPostsByUser();
+          } else {
+            this.showPosts();
+          }
+          this.loadPendingPosts();
+        }, 500);
+      });
     });
   }
 
@@ -283,10 +362,12 @@ export class FreedomWallComponent implements OnInit {
   getDeletionRequests(){
     this.dataService.getDeletionRequests().subscribe((posts:Post[])=>{
       this.requestDelete =posts;
+      this.updateButtonCounts();
     });
   }
 
   requestPostToDelete(post: Post){
+    this.confirmationDialogService.confirmAction('Request Confirmation', 'This action cant be undone. Are you sure you want to request to delete this post?', () => {
     const post_id = { id: post.id };
     this.dataService.deletionRequest(post_id).subscribe((res: Response)=>{
       this.response = res;
@@ -306,7 +387,8 @@ export class FreedomWallComponent implements OnInit {
       this.toggleOptions(post);
       this.getDeletionRequests();
     });
-  }
+  });
+}
 
   declineRequestToDelete(id: number){
     const post_id = { id: id };
@@ -327,6 +409,37 @@ export class FreedomWallComponent implements OnInit {
       }
       this.getDeletionRequests();
     })
+  }
 
+  updateButtonCounts(): void {
+    this.deletePostCount = this.requestDelete ? this.requestDelete.length : 0;
+    this.manageWallCount = this.pendingPosts ? this.pendingPosts.length : 0;
+  }
+
+  onDropdownClick(event: MouseEvent) {
+    event.stopPropagation();
+  }
+
+  filterPostsByUser(): Post[] {
+    this.showFilterMessage = true;
+    this.showSpinner = true; 
+    this.reloadMasonryLayout(); 
+    setTimeout(() => {
+      this.showSpinner = false;
+
+    }, 1000); 
+    return this.posts.filter(post => post.student_id === this.userInfo.id);
+    
+  }
+
+  clearFilter(): void {
+    this.showFilterMessage = false;
+    this.showSpinner = true; 
+    this.showPosts(); 
+  
+    setTimeout(() => {
+      this.reloadMasonryLayout(); 
+      this.showSpinner = false;
+    }, 1000); 
   }
 }
